More refactoring and did some changes

+++ app/engine/objects/object.hpp
@@ -17,7 +17,7 @@ enum {
 };
 
 struct TransformComponent {
-  glm::vec3 translation{};  // (position offset)
+  glm::dvec3 translation{};  // (position offset)
   glm::vec3 scale{1.f, 1.f, 1.f};
   glm::vec3 rotation{};
